+++ src/main/java/ua/edu/ucu/stream/IntStream.java
@@ -1,14 +1,18 @@
 package ua.edu.ucu.stream;
 
-import ua.edu.ucu.function.*;
+import ua.edu.ucu.function.IntToIntStreamFunction;
+import ua.edu.ucu.function.IntPredicate;
+import ua.edu.ucu.function.IntConsumer;
+import ua.edu.ucu.function.IntUnaryOperator;
+import ua.edu.ucu.function.IntBinaryOperator;
 
 public interface IntStream {
 
-    Double average();
+    long average();
 
-    Integer max();
+    int max();
 
-    Integer min();
+    int min();
     
     IntStream flatMap(IntToIntStreamFunction func);
